Docs are stackifyed

+++ test/unit-testing/BaseSpec.hs
@@ -43,35 +43,26 @@ spec = do
   
     describe "dot" $ do
       it "pair" $ dot @XY @XY (1, 2) (3, 6) == 15
-      it "complex" $ dot @Cmp
-       @Cmp
-       (1 :+ 2) (3 :+ 6) == 15
+      it "complex" $ dot @Cmp @Cmp (1 :+ 2) (3 :+ 6) == 15
   
     describe "cross" $ do
       it "pair" $ cross @XY @XY (1, 2) (3, 6) == 0
-      it "complex" $ cross @Cmp
-       @Cmp
-       (1 :+ 2) (3 :+ 6) == 0
+      it "complex" $ cross @Cmp @Cmp (1 :+ 2) (3 :+ 6) == 0
     
     describe "norm" $ do
       it "pair" $ norm @XY (3, 4) == 5
-      it "complex" $ norm @Cmp
-       (3 :+ 4) == 5
+      it "complex" $ norm @Cmp (3 :+ 4) == 5
 
     describe "normalize" $ do
       it "pair" $ norm (normalize @XY (3, 4)) == 1
-      it "complex" $ norm (normalize @Cmp
-       (3 :+ 4)) == 1
-      it "trivial" $ normalize @Cmp
-       0 == 0
+      it "complex" $ norm (normalize @Cmp (3 :+ 4)) == 1
+      it "trivial" $ normalize @Cmp 0 == 0
 
   describe "Transformations" $ do
     describe "translate" $ do
       it "1" $ translate' @XY @XY (10, 20) (3, 4)  == (13, 24)
-      it "2" $ translate' @XY @Cmp
-       10 (3, 4)  == (13, 4)
-      it "3" $ translate' @Cmp
-       @XY (3, 4) 10  == 13 :+ 4
+      it "2" $ translate' @XY @Cmp 10 (3, 4)  == (13, 4)
+      it "3" $ translate' @Cmp @XY (3, 4) 10  == 13 :+ 4
 
     it "scale" $ scale @XY 10 (3, 4)  == (30, 40)
     it "scaleX" $ scaleX @XY 10 (3, 4) ~= (30, 4)
@@ -95,14 +86,6 @@ spec = do
       it "3" $ rotateAt' @XY @XY (1, 0) 90 (0, 0) ~= (1, -1)
       it "4" $ rotateAt' @XY @XY (1, 1) 180 (0, 0) ~= (2, 2)
 
-    -- describe "orientation" $ do
-    --   it "1" $ property $ \a -> transformOrientation (rotateT a) == 1
-    --   it "2" $ property $ \v -> transformOrientation (translateT v) == 1
-    --   it "3" $ property $ \sx sy ->
-    --     transformOrientation (scaleT sx sy) == signum sx * signum sy
-    --   it "4" $ property $ \a ->
-    --     transformOrientation (reflectT a) == -1
-
   describe "Box" $ do
     let l = Segment (1:+2, 4:+6)
     it "0" $ box l == ((1,2),(4,6))
@@ -116,7 +99,6 @@ spec = do
   describe "AlmostEqual" $ do
     it "1" $ property $ \i j -> let _ = (i :: Int, j :: Int)
                                 in (i == j) == (i ~= j)
-    it "2" $ property $ \i j -> let _ = (i :: Maybe Double, j :: Maybe Double)
+    it "2" $ property $ \i j -> let _ = (i :: Maybe Int, j :: Maybe Int)
                                 in (i == j) == (i ~= j)
---    it "3" $ property $ \i j -> let _ = (i :: Double, j :: Double)
---                                in (i == j) == (i ~= j)
+
